Remove draft-ietf-ppm-dap-02
Drop support for DAP draft 02 and prune some code that is no longer
needed:

1. Remove `AggregationJobContinueReq`, as this is only needed to support
   multi-round VDAFs. For now we only support 1-round VDAFs.

2. Remove code for draft-wang-ppm-dap-taskprov-02, including processing
   of `TaskConfig` struct carried by report extensions.

2. Remove `MetaAggregationJobId`. This was used to unify shared
   semantics across draft 02 and 09, which have a different syntax for
   aggregation job IDs.

3. Remove the code in draft 02 for generating a collection job ID.

As a result of dropping draft 02, it is possible to further simplify
control flow and some protocol semantics (e.g., returning multiple HPKE
configs). To keep the commit as small as possible, we leave most of
these changes as TODOs.