draft02 cleanup: Simplify state machine
In response to `AggregationJobInitReq`: in draft02 the Helper would wait
for one more message before computing its agg share span; in draft09 it
computes its agg share span immediately. The following types were
introduced to deal with this complexity:

DapLeaderAggregationJobTransition
DapHelperAggregationJobTransition

Now that we've removed draft02, we can replace with explicit types and
remove the code for handling unexpected transitions.

The new state machine is simply:

- Leader calls `produce_agg_job_req()` to get its
  `DapAggregationJobState` and the outbound request

- Helper calls `consume_agg_job_req()` to get the initialize reports
  `Vec<EarlyReportStateInitialized>`

- Helper calls `produce_agg_job_resp()` to get the outbound response and
  its agg share span

- Leader calls `consume_agg_job_resp()` to get its agg share span