Rework durable object definitions
Remove all the trait complexitiy in durable.rs and reduce it all to one
trait and one macro.

The DapDurableObject trait defines the minimum set of methods required
for the macro to be able to define a durable object struct.

Delete the DaphneWorkerDurableConfig struct. All of the values from this
struct are defined at deploy time from environment variables and as such
can all be cached as static init once variables that can be scoped the
place in which their used:

- deployment was moved to a function that loads the deployment mode.
- the helper gc time was moved to the should_cleanup_at function.

the aggregate store object now has gc enabled with a default of one week.

Behavior changes: Self gc from durable object instances is now counted
from last touch instead of first touch.