Move report time validity range check to protocol
`daphne_server::App` currently implements a method
`valid_report_time_range()` used to check if a report is too far in the
future or too far in the past to be processed. This is currently called
during report initialization. We make two observations:

1. This check is going to be needed by any implementation of the rele
   traits, not just `daphne_server::App`.

2. This check can be done much earlier, even before decryption.

To address these points, we make `valid_report_time_range()` a method on
`DapReportInitializer` and call it during report consumption. While at
it, remove the unused `early_metadata_check()` function, which is no
longer used.