Fix bug in `try_put_agg_share_span()` retry loop
First remove the `Finished` variant of `TransitionVar` as its not needed
for 1-round VDAFs.

After making this change, we notice a bug in the
`try_put_agg_share_span()` retry loop: if a report was aggregated during
the previous attempt (its `ReportProcessedStatus` is `Aggregated`), then
we would have returned `TransitionVar::Finished`, which is incorrect. If
a report was aggregated, we instead want to compute the `outbound`
message and return `TransitionVar::Continued(outbound)`.