pine: Specify number of proofs and algorithm ID dynamically
Currently the number of FLPs to generate is a generic constant and thus
determined at compile time. For taskprov, the number of proofs will be
configured at runtime. To make this easier, move this parameter into
`PineConfig`.

This change also moves the VDAF algorithm ID to `PineConfig`. This is
needed for a different reason however: we will have more than one
instance of PINE, for different field sizes and XOFs.