pine:  Represent norm bound as integer rather than float
Change the type of the `norm_bound` parameter of `Pine::new()` from
`f64` and `u64`. The caller is expected to compute this as `floor(b *
2^frac_bits)`, where `b: f64` is the real-valued, L2-norm bound.

All parties must agree on the algorithm used to derive Pine's internal
parameters from `norm_bound` parameter. Representing it this way is
intended to the complexity of this algorithm.

Corresponding changes to the spec:
https://github.com/junyechen1996/draft-chen-cfrg-vdaf-pine/pull/87
https://github.com/junyechen1996/draft-chen-cfrg-vdaf-pine/pull/88