storage: add undocumented environment variable for location hinting
For some of our uses at Cloudflare, it would be useful to deploy instances of Daphne which use location hints to pin DOs to specific DO regions.

This change adds this support as an undocumented build-time environment variable, DAPHNE_DO_REGION.

We can eventually work out whether we want this to be exposed in a different way or not, and then either document this environment variable and stabilize it or expose this data to the worker another way.