pine: Compute the squared norm just once
Split the existing FLP circuit into two:

1. One that computes the squared norm bound and checks that it is equal
   to the claimed value

2. Another for everything else, including checking that the claimed norm
   bound is in range and that the wraparound tests succeeded

The first circuit does not require joint randomness to compute, which
means we can safely run it once without opening ourselves to offline
attacks. It is also the most expensive part of the computation.

This change aligns our implementation with a planned change for the next
version of PINE:

https://github.com/junyechen1996/draft-chen-cfrg-vdaf-pine/pull/92
https://github.com/junyechen1996/draft-chen-cfrg-vdaf-pine/pull/94