pine: Add additional test vectors and fix a bug
Add additional test vectors for PINE. While at it, make the test runner
a one-liner that can be more easily copy+pasted.

The new vectors revealed a bug in the `eval_wr_checks()` sub-circuit
related to improper handling of the last chunk of inputs to the gadget.

The bug is an incorrect order of operations that was triggered by the
number of wraparound tests used by the new test vectors. The last chunk
is skipped if the condition `buf.len() % self.cfg.param.chunk_len * 2 !=
0` holds, where `buf` is the last chunk. In our case, `buf.len()` is 150
(2 times 75 wraparound tests) and `self.cfg.param.chunk_len` is 150. The
expression gets evaluated as `(150 % 150) * 2 != 0`, not `150 % (150 *
2) != 0`, and thus the last chunk gets skipped.

The correct condition is `!buf.is_empty()`.