pine: Avoid duplicating unused parameters
Internally, `Pine::new()` constructs a `PineConfig` that is used to
construct both FLPs (the squared norm equality check and the main
check). This means that some parameters are copied without ever being
used. It's better to copy only what we need.

While at it, rename `gadget_calls_sq_norm_equal` to `gadget_calls` now
that we don't need to disambiguate this parameter across the two FLPs.
This is better aligned with the PINE spec.