Unify VDAF dispatching logic
Add dispatchers for sharding, preparation, and unsharding to
`VdafConfig`. Pass the DAP version to the dispatcher so that we can
properly resolve supported VDAFs.

Remove support for the following, unspecified combinations:

* DAP-13/Pine32HmacSha256Aes128
* DAP-13/Pine64HmacSha256Aes128
* DAP-09/Mastic

Technically, since Mastic is still a shim at the moment and not a
functional VDAF, we could support it in DAP-09 as well. When the
implementation is complete
(https://github.com/divviup/libprio-rs/issues/947), it will only be
VDAF-13-compatible and thus only supported in DAP-13.